driver: restructure image acquisition thread to not miss frames
Previously only the latest image was read out, so if images came in
faster than the driver responded with events images were missed.

We now track the frame stamp (incrementing frame index) and read
out all new frames each event, meaning that we should not miss
frames, and that if we do we know about it!